+++ tests/TestsUtil/Util.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace SugarCli\Tests\TestsUtil;
+
+use Symfony\Component\Filesystem\Filesystem;
+
+class Util
+{
+    public static function getRelativePath($path)
+    {
+        $fs = new Filesystem();
+        return $fs->makePathRelative(
+            $path,
+            getcwd()
+        );
+    }
+}